03-botcity/bot-web/bot-web/requirements.txt
# Uncomment to use botcity-utils
# botcity-utils>=0.1.2

# Uncomment to use botcity-crawler-plugin
# botcity-crawler-plugin

# Uncomment to use Pandas
# pandas

botcity-framework-web>=1.0,<2.0
botcity-maestro-sdk>=0.5.1,<1.0
chromedriver
selenium==4.10.0
webdriver-manager==4.0.2
